Got rid of sand traps and water
For now hole represents enemy as I couldn'y firgure out how to change the hole's sprite. Also there is currently drag on the bullet that slows it down but to fix this delete line 316 when you figure out how to stop ball movement after three bounces because I couldn't do that part. You also still need to add barriers into the levels, a timer, and a certain amount of ammo.